+++ test/cases/require-tmpl-in-js/src/app.js
@@ -1,0 +1,4 @@
+import './component-cjs/index';
+import './component-esm/index';
+
+console.log('>> app');