+++ test/cases/require-tmpl-in-js/src/component-esm/index.js
@@ -1,0 +1,4 @@
+import tmpl from './template.html?name=ES Module';
+
+const appContainer = document.getElementById('app-esm');
+appContainer.innerHTML = tmpl;