+++ test/cases/require-tmpl-in-js/webpack.config.js
@@ -1,0 +1,38 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  entry: {
+    index: './src/index.html',
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      loaderOptions: {
+        data: {
+          component: {
+            title: 'My component',
+          },
+        },
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(ico|png|jp?g|svg)$/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'img/[name].[hash:8][ext][query]',
+        },
+      },
+    ],
+  },
+};