+++ test/jest.config.js
@@ -205,7 +205,7 @@ module.exports = {
   // testSequencer: '@jest/test-sequencer',
 
   // Default timeout of a test in milliseconds.
-  testTimeout: isLocalEnv ? 1000 : 10000,
+  testTimeout: isLocalEnv ? 1000 : 5000,
 
   // This option sets the URL for the jsdom environment. It is reflected in properties such as location.href
   // testURL: "http://localhost",