+++ test/manual/tmpl-in-js-publicPath-auto/src/views/main.js
@@ -1,0 +1,2 @@
+import './component/index';
+console.log('>> main');